+++ packages/presets/src/presets/preact/index.ts
@@ -1,0 +1,4 @@
+import { AzionBuildPreset } from '../../types';
+import config from './config';
+const preset: AzionBuildPreset = { config };
+export default preset;