+++ packages/sql/src/index.test.ts
@@ -11,17 +11,24 @@ import {
 import * as servicesApi from '../src/services/api/index';
 import * as services from '../src/services/index';
 import { AzionSQLClient } from './types';
+import fetchWithErrorHandling from './utils/fetch';
 
 jest.mock('../src/services/api/index');
+jest.mock('../src/utils/fetch/index', () => ({
+  __esModule: true,
+  default: jest.fn(),
+}));
 
 describe('SQL Module', () => {
   const mockToken = 'mock-token';
   const mockDebug = true;
 
   beforeEach(() => {
-    jest.clearAllMocks();
+    jest.resetAllMocks();
     process.env.AZION_TOKEN = mockToken;
     process.env.AZION_DEBUG = 'true';
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    (globalThis as any).Azion = {};
   });
 
   describe('createClient', () => {
@@ -51,7 +58,7 @@ describe('SQL Module', () => {
       (servicesApi.postEdgeDatabase as jest.Mock).mockResolvedValue(mockResponse);
       const result = await createDatabase('test-db', { debug: mockDebug });
       expect(result.data).toEqual(expect.objectContaining({ id: 1, name: 'test-db' }));
-      expect(servicesApi.postEdgeDatabase).toHaveBeenCalledWith(mockToken, 'test-db', true);
+      expect(servicesApi.postEdgeDatabase).toHaveBeenCalledWith(mockToken, 'test-db', true, 'production');
     });
 
     it('should return error on failure', async () => {
@@ -76,7 +83,7 @@ describe('SQL Module', () => {
       (servicesApi.deleteEdgeDatabase as jest.Mock).mockResolvedValue({ state: 'pending', data: { id: 1 } });
       const result = await deleteDatabase(1, { debug: mockDebug });
       expect(result).toEqual({ data: { id: 1, state: 'pending' } });
-      expect(servicesApi.deleteEdgeDatabase).toHaveBeenCalledWith(mockToken, 1, true);
+      expect(servicesApi.deleteEdgeDatabase).toHaveBeenCalledWith(mockToken, 1, true, 'production');
     });
 
     it('should return error on failure', async () => {
@@ -100,7 +107,7 @@ describe('SQL Module', () => {
 
       const result = await getDatabase('test-db', { debug: mockDebug });
       expect(result.data).toEqual(expect.objectContaining({ id: 1, name: 'test-db' }));
-      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith(mockToken, { search: 'test-db' }, true);
+      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith(mockToken, { search: 'test-db' }, true, 'production');
     });
 
     it('should throw an error if database is not found', async () => {
@@ -128,7 +135,12 @@ describe('SQL Module', () => {
       const { data } = await getDatabases({ page: 1, page_size: 10 }, { debug: mockDebug });
       expect(data?.databases).toHaveLength(2);
       expect(data!.databases![0]).toEqual(expect.objectContaining({ id: 1, name: 'test-db-1' }));
-      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith(mockToken, { page: 1, page_size: 10 }, true);
+      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith(
+        mockToken,
+        { page: 1, page_size: 10 },
+        true,
+        'production',
+      );
     });
 
     it('should return error if retrieval fails', async () => {
@@ -273,11 +285,12 @@ describe('SQL Module', () => {
     it('should successfully execute useQuery with apiQuery called', async () => {
       const spyApiQuery = jest.spyOn(services, 'apiQuery').mockResolvedValue({ data: { toObject: () => null } });
 
-      await useQuery('test-db', ['SELECT * FROM test'], { debug: mockDebug });
+      await useQuery('test-db', ['SELECT * FROM test'], { debug: mockDebug, env: 'production' });
 
       expect(spyApiQuery).toHaveBeenCalled();
       expect(spyApiQuery).toHaveBeenCalledWith(mockToken, 'test-db', ['SELECT * FROM test'], {
         debug: mockDebug,
+        env: 'production',
       });
 
       spyApiQuery.mockRestore();
@@ -295,9 +308,156 @@ describe('SQL Module', () => {
       expect(spyRuntimeQuery).toHaveBeenCalled();
       expect(spyRuntimeQuery).toHaveBeenCalledWith(mockToken, 'test-db', ['SELECT * FROM test'], {
         debug: mockDebug,
+        env: 'production',
       });
 
       spyRuntimeQuery.mockRestore();
+    });
+
+    it('should return error if useQuery when statement some invalid', async () => {
+      const mockResponseDatabases = {
+        results: [
+          { id: 1, name: 'test-db' },
+          { id: 2, name: 'test-db-2' },
+        ],
+      };
+      (servicesApi.getEdgeDatabases as jest.Mock).mockResolvedValue(mockResponseDatabases);
+
+      (servicesApi.postQueryEdgeDatabase as jest.Mock).mockImplementation(
+        jest.requireActual('../src/services/api/index').postQueryEdgeDatabase,
+      );
+
+      (fetchWithErrorHandling as jest.Mock).mockResolvedValue({
+        state: 'executed',
+        data: [
+          {
+            results: {
+              columns: ['schema', 'name', 'type', 'ncol', 'wr', 'strict'],
+              rows: [
+                ['main', 'sqlite_schema', 'table', 5, 0, 0],
+                ['temp', 'sqlite_temp_schema', 'table', 5, 0, 0],
+              ],
+              rows_read: 0,
+              rows_written: 0,
+              query_duration_ms: 0.058,
+            },
+          },
+          {
+            error: 'no such table: main',
+          },
+          null,
+        ],
+      });
+
+      await expect(
+        useQuery('test-db', ['pragma table_list', 'select * from main', 'select * from sqlite_schema'], {
+          debug: mockDebug,
+        }),
+      ).resolves.toEqual(
+        expect.objectContaining({
+          error: {
+            message: 'no such table: main',
+            operation: 'post query',
+          },
+        }),
+      );
+    });
+
+    it('should return error if useQuery when last statement is invalid', async () => {
+      const mockResponseDatabases = {
+        results: [
+          { id: 1, name: 'test-db' },
+          { id: 2, name: 'test-db-2' },
+        ],
+      };
+      (servicesApi.getEdgeDatabases as jest.Mock).mockResolvedValue(mockResponseDatabases);
+
+      (servicesApi.postQueryEdgeDatabase as jest.Mock).mockImplementation(
+        jest.requireActual('../src/services/api/index').postQueryEdgeDatabase,
+      );
+
+      (fetchWithErrorHandling as jest.Mock).mockResolvedValue({
+        state: 'executed',
+        data: [
+          {
+            results: {
+              columns: ['schema', 'name', 'type', 'ncol', 'wr', 'strict'],
+              rows: [
+                ['main', 'sqlite_schema', 'table', 5, 0, 0],
+                ['temp', 'sqlite_temp_schema', 'table', 5, 0, 0],
+              ],
+              rows_read: 0,
+              rows_written: 0,
+              query_duration_ms: 0.058,
+            },
+          },
+          {
+            results: { columns: ['id', 'name'], rows: [[1, 'test']] },
+          },
+          {
+            error: 'no such table: main',
+          },
+        ],
+      });
+
+      await expect(
+        useQuery('test-db', ['pragma table_list', 'select * from main', 'select * from sqlite_schema'], {
+          debug: mockDebug,
+        }),
+      ).resolves.toEqual(
+        expect.objectContaining({
+          data: expect.objectContaining({
+            state: 'executed',
+            results: [
+              {
+                columns: ['schema', 'name', 'type', 'ncol', 'wr', 'strict'],
+                rows: [
+                  ['main', 'sqlite_schema', 'table', 5, 0, 0],
+                  ['temp', 'sqlite_temp_schema', 'table', 5, 0, 0],
+                ],
+                statement: 'pragma',
+              },
+              { columns: ['id', 'name'], rows: [[1, 'test']], statement: 'select' },
+            ],
+          }),
+          error: {
+            message: 'no such table: main',
+            operation: 'post query',
+          },
+        }),
+      );
+    });
+
+    it('should return error if useQuery when data rows is empty', async () => {
+      const mockResponseDatabases = {
+        results: [
+          { id: 1, name: 'test-db' },
+          { id: 2, name: 'test-db-2' },
+        ],
+      };
+      (servicesApi.getEdgeDatabases as jest.Mock).mockResolvedValue(mockResponseDatabases);
+
+      (servicesApi.postQueryEdgeDatabase as jest.Mock).mockImplementation(
+        jest.requireActual('../src/services/api/index').postQueryEdgeDatabase,
+      );
+
+      (fetchWithErrorHandling as jest.Mock).mockResolvedValue({
+        state: 'executed',
+        data: [{ results: { columns: ['id', 'name'], rows: [] } }],
+      });
+
+      await expect(
+        useQuery('test-db', ['select * from users'], {
+          debug: mockDebug,
+        }),
+      ).resolves.toEqual(
+        expect.objectContaining({
+          data: expect.objectContaining({
+            state: 'executed',
+            results: [{ columns: ['id', 'name'], rows: undefined, statement: 'select' }],
+          }),
+        }),
+      );
     });
   });
 
@@ -354,7 +514,7 @@ describe('SQL Module', () => {
       (servicesApi.postEdgeDatabase as jest.Mock).mockResolvedValue(mockResponse);
 
       await client.createDatabase('test-db');
-      expect(servicesApi.postEdgeDatabase).toHaveBeenCalledWith('custom-token', 'test-db', false);
+      expect(servicesApi.postEdgeDatabase).toHaveBeenCalledWith('custom-token', 'test-db', false, 'production');
     });
 
     it('should call deleteDatabase method', async () => {
@@ -362,7 +522,7 @@ describe('SQL Module', () => {
       (servicesApi.deleteEdgeDatabase as jest.Mock).mockResolvedValue(mockResponse);
 
       await client.deleteDatabase(1);
-      expect(servicesApi.deleteEdgeDatabase).toHaveBeenCalledWith('custom-token', 1, false);
+      expect(servicesApi.deleteEdgeDatabase).toHaveBeenCalledWith('custom-token', 1, false, 'production');
     });
 
     it('should call getDatabase method', async () => {
@@ -371,7 +531,12 @@ describe('SQL Module', () => {
       if (client.getDatabase) {
         await client.getDatabase('test-db');
       }
-      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith('custom-token', { search: 'test-db' }, false);
+      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith(
+        'custom-token',
+        { search: 'test-db' },
+        false,
+        'production',
+      );
     });
 
     it('should call getDatabases method', async () => {
@@ -384,7 +549,12 @@ describe('SQL Module', () => {
       (servicesApi.getEdgeDatabases as jest.Mock).mockResolvedValue(mockResponse);
 
       await client.getDatabases({ page: 1, page_size: 10 });
-      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith('custom-token', { page: 1, page_size: 10 }, false);
+      expect(servicesApi.getEdgeDatabases).toHaveBeenCalledWith(
+        'custom-token',
+        { page: 1, page_size: 10 },
+        false,
+        'production',
+      );
     });
   });
 });